Function northstar_client::connect

pub async fn connect<T: AsyncRead + AsyncWrite + Unpin>(
    io: T, 
    subscribe_notifications: bool
) -> Result<Connection<T>, ConnectionError>

Connect and return a raw stream and sink interface. See codec for details

Arguments

io - Medium for the connection (e.g. Unix or TCP socket)

subscribe_notifications - Enables the reception of notifications through the connection.

Errors

An error is returned in the following cases:

A mismatch in the protocol version between both sides of the connection

Unnecessary permissions

OS errors