Struct northstar_client::Client

pub struct Client<T> { /* private fields */ }

Client for a Northstar runtime instance.

use futures::StreamExt;
use tokio::time::Duration;
use northstar_client::Client;
use northstar_client::model::Version;

async fn main() {
    let mut client = Client::new(tokio::net::TcpStream::connect("localhost:4200").await.unwrap(), None, Duration::from_secs(10)).await.unwrap();
    client.start("hello:0.0.1").await.expect("failed to start \"hello\"");
    while let Some(notification) = client.next().await {
        println!("{:?}", notification);
    }
}

Implementations

impl Client<UnixStream>

pub async fn from_env(
notifications: Option<usize>,
timeout: Duration
) -> Result<Self, FromEnvError>

Tries to create a client by accessing NORTHSTAR_CONSOLE env variable

Errors

An Err is returned if the NORTHSTAR_CONSOLE environment variable is not set or has an invalid file descriptor for the unix socket.

impl<'a, T: AsyncRead + AsyncWrite + Unpin> Client<T>

pub async fn new(
io: T,
notifications: Option<usize>,
timeout: Duration
) -> Result<Client<T>, ClientError>

Create a new northstar client and connect to a runtime instance running on host.

Arguments

io - Connection medium (e.g. Unix or TCP socket)

notifications - Optional buffer size for receiving notifications

timeout - Timeout of connection establishment

Errors

In addition to the errors that can happen when trying to connect, an Err is returned if the connection establishment times out.

pub fn framed(self) -> Connection<T>

Convert client into a connection

pub async fn request(
&mut self,
request: Request
) -> Result<Response, RequestError>

Perform a request response sequence

let response = client.request(List).await.expect("failed to request container list");
println!("{:?}", response);

pub async fn ident(&mut self) -> Result<Container, RequestError>

Request the identificaiton of this container

let ident = client.ident().await.expect("failed to identity");
println!("{}", ident);

pub async fn list(&mut self) -> Result<Vec<Container>, RequestError>

Request a list of installed containers

let containers = client.list().await.expect("failed to request container list");
println!("{:#?}", containers);

pub async fn repositories(
&mut self
) -> Result<HashSet<RepositoryId>, RequestError>

Request a list of repositories

let repositories = client.repositories().await.expect("failed to request repository list");
println!("{:#?}", repositories);

pub async fn start<C>(&mut self, container: C) -> Result<(), RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,

Start container with name

client.start("hello:0.0.1").await.expect("failed to start \"hello\"");
// Print start notification
println!("{:#?}", client.next().await);

pub async fn start_with_args<C, A>(
&mut self,
container: C,
args: impl IntoIterator<Item = A>
) -> Result<(), RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,
A: TryInto<NonNulString>,
A::Error: Error + Send + Sync + 'static,

Start container name and pass args

client.start_with_args("hello:0.0.1", ["--foo"]).await.expect("failed to start \"hello --foor\"");
// Print start notification
println!("{:#?}", client.next().await);

pub async fn start_with_args_env<C, A>(
&mut self,
container: C,
args: impl IntoIterator<Item = A>,
env: impl IntoIterator<Item = (A, A)>
) -> Result<(), RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,
A: TryInto<NonNulString>,
A::Error: Error + Send + Sync + 'static,

Start container name and pass args and set additional env variables

let mut env = HashMap::new();
env.insert("FOO", "blah");
client.start_with_args_env("hello:0.0.1", ["--dump", "-v"], env).await.expect("failed to start \"hello\"");
// Print start notification
println!("{:#?}", client.next().await);

pub async fn kill<C>(
&mut self,
container: C,
signal: i32
) -> Result<(), RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,

Kill container with name

client.kill("hello:0.0.1", 15).await.expect("failed to start \"hello\"");
// Print stop notification
println!("{:#?}", client.next().await);

pub async fn install_file(
&mut self,
npk: &Path,
repository: &str
) -> Result<Container, RequestError>

Install a npk from path

let npk = Path::new("test.npk");
client.install_file(npk, "default").await.expect("failed to install \"test.npk\" into repository \"default\"");

pub async fn install(
&mut self,
npk: impl AsyncRead + Unpin,
size: u64,
repository: &str
) -> Result<Container, RequestError>

Install a npk

let npk = fs::File::open("test.npk").await.expect("failed to open \"test.npk\"");
let size = npk.metadata().await.unwrap().len();
client.install(npk, size, "default").await.expect("failed to install \"test.npk\" into repository \"default\"");

pub async fn uninstall<C>(
&mut self,
container: C,
wipe: bool
) -> Result<(), RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,

Uninstall a npk and optionally wipe the containers persistent dir

client.uninstall("hello:0.0.1", false).await.expect("failed to uninstall \"hello\"");
// Print stop notification
println!("{:#?}", client.next().await);

pub async fn shutdown(&mut self)

Stop the runtime

pub async fn mount<C>(
&mut self,
container: C
) -> Result<MountResult, RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,

Mount a container

let mut client = Client::new(tokio::net::TcpStream::connect("localhost:4200").await.unwrap(), None, Duration::from_secs(10)).await.unwrap();
client.mount("test:0.0.1").await.expect("failed to mount");

pub async fn mount_all<C, I>(
&mut self,
containers: I
) -> Result<Vec<MountResult>, RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,
I: 'a + IntoIterator<Item = C>,

Mount a list of containers

let mut client = Client::new(tokio::net::TcpStream::connect("localhost:4200").await.unwrap(), None, Duration::from_secs(10)).await.unwrap();
client.mount_all(vec!("hello-world:0.0.1", "cpueater:0.0.1")).await.expect("failed to mount");

pub async fn umount<C>(
&mut self,
container: C
) -> Result<UmountResult, RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,

Umount a mounted container

client.umount("hello:0.0.1").await.expect("failed to unmount \"hello:0.0.1\"");

pub async fn umount_all<C, I>(
&mut self,
containers: I
) -> Result<Vec<UmountResult>, RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,
I: 'a + IntoIterator<Item = C>,

Umount a list of mounted containers

client.umount_all(vec!("hello:0.0.1", "cpueater:0.0.1")).await.expect("failed to unmount \"hello:0.0.1\" and \"cpueater:0.0.1\"");

pub async fn inspect<C>(
&mut self,
container: C
) -> Result<ContainerData, RequestError> where
C: TryInto<Container>,
C::Error: Error + Send + Sync + 'static,

Gather container statistics

println!("{:?}", client.inspect("hello:0.0.1").await.unwrap());

pub async fn create_token<R, S>(
&mut self,
target: R,
shared: S
) -> Result<Token, RequestError> where
R: TryInto<Name>,
R::Error: Error + Send + Sync + 'static,
S: AsRef<[u8]>,

Create a token

The target parameter must be the container name (without version) of the container that will try to verify the token. The token can only be successfully verified by the container that is started with the name target! The shared parameter is added into the token in order to make it specific to a dedicated purpose, e.g. “mqtt”.

println!("{:?}", client.create_token("webserver", "http").await.unwrap());

pub async fn verify_token<R, S>(
&mut self,
token: &Token,
user: R,
shared: S
) -> Result<VerificationResult, RequestError> where
R: TryInto<Name>,
R::Error: Error + Send + Sync + 'static,
S: AsRef<[u8]>,

Verify a slice of bytes with a token

The token parameter shall contain a token that is received from a container. The user parameter must match the name of the container, that created the token and send it to the container that want to verify the token. shared is some salt that makes a token specific for a usecase can must just match the value used when the the token is created.

let token = client.create_token("hello", "#noafd").await.unwrap(); // token can only verified by container `hello`
assert_eq!(client.verify_token(&token, "hello", "#noafd").await.unwrap(), VerificationResult::Ok);
assert_eq!(client.verify_token(&token, "hello", "").await.unwrap(), VerificationResult::Ok);